+++ src/app/(PianoRoll)/PianoRoll.tsx
@@ -10,8 +10,9 @@ import PianoRollRow from "./PianoRow";
 import getGridDimensions from "./getGridDimensions";
 import { DragSelectProvider } from "./DragSelectProvider";
 import { DSInputElement } from "dragselect";
-import useDragResize from "./useDragResize";
+import useResizeHandlers from "./useResizeHandlers";
 import { useNoteHandlers } from "./useNoteHandlers";
+import useDragHandlers from "./useDragHandlers";
 
 export function widthFromDurationTicks(ticks: number, cellWidth: number) {
     return (ticks / TICKS_PER_16TH) * cellWidth;
@@ -23,16 +24,6 @@ export function ticksFromWidth(width: number, cellWidth: number) {
 
 // TODO(will): This should be a user setting. Also maybe can add more fine-grained control over quantization.
 const QUANTIZED = true;
-
-function getNoteFromId(id: string, notes: ExtendedNoteJSON[]) {
-    const [_, _1, noteName, ticks, durationTicks] = id.split("-");
-    return notes.find(
-        (n) =>
-            n.name === noteName &&
-            n.ticks === parseInt(ticks) &&
-            n.durationTicks === parseInt(durationTicks)
-    );
-}
 
 export function PianoRoll({
     notes,
@@ -56,7 +47,7 @@ export function PianoRoll({
     const { cellHeight, cellWidth, totalColumns, totalWidth } =
         getGridDimensions(notes, width, height);
 
-    const { commitResize, handleResize, resizeWidth } = useDragResize(
+    const { commitResize, handleResize, resizeWidth } = useResizeHandlers(
         selectedNotes,
         setSelectedNotes,
         cellWidth,
@@ -73,66 +64,23 @@ export function PianoRoll({
         notes
     );
 
-    const updateNotes = ({
-        items,
-        event,
-    }: {
-        items: DSInputElement[];
-        event?: MouseEvent | TouchEvent | null | undefined | KeyboardEvent;
-    }) => {
-        if (!items.length) return;
-
-        const noteIds = items.map((item) => item.id);
-        const notesFromIds = noteIds.map((id) => {
-            return getNoteFromId(id, notes);
-        });
-
-        const transform = items[0].style.transform;
-        // No transform means we are not dragging
-        if (!transform) return;
-
-        // Get the transform values
-        const transformX = transform.split("(")[1].split(",")[0];
-        const transformY = transform.split("(")[1].split(",")[1];
-
-        // Remove the old selected notes
-        removeNotes(notesFromIds.filter((n) => n !== undefined));
-
-        // Create new, relocated notes
-        const notesToAdd = notesFromIds
-            .filter((n) => n !== undefined)
-            .map((n) =>
-                deriveNewNote(
-                    parseInt(transformX),
-                    parseInt(transformY),
-                    cellHeight,
-                    n,
-                    QUANTIZED
-                )
-            )
-            .filter((n) => n !== null);
-
-        // Add the new notes
-        addNotes(notesToAdd);
-        // Set the selected notes to the new notes
-        setSelectedNotes(notesToAdd);
-    };
-
-    const isDraggable = mode === "point";
+    const { handleDragEnd, handleDragSelect, isDraggable } = useDragHandlers(
+        mode,
+        notes,
+        cellHeight,
+        QUANTIZED,
+        addNotes,
+        removeNotes,
+        setSelectedNotes
+    );
 
     return (
         <DragSelectProvider
             settings={{
                 area: dragContainer.current ?? undefined,
             }}
-            onDragEnd={updateNotes}
-            onDragSelect={(items: DSInputElement[]) => {
-                setSelectedNotes(
-                    items
-                        .map((item) => getNoteFromId(item.id, notes))
-                        .filter((n) => n !== undefined)
-                );
-            }}
+            handleDragEnd={handleDragEnd}
+            handleDragSelect={handleDragSelect}
         >
             <div
                 className="relative border border-zinc-700 overflow-auto"
